Change Coin data structure, refactor VendingMachine
Use 'properties' and 'types' for coins.
A coin's properties are its weight and diameter.
A coin type consists of a property and value.
Additional properties can be added if desired.

Helper static methods added to Coin to sort an array of coins by value
and return the value of an array of coins.

Refactor concerns in VendingMachine select and makeChange.
MakeChange only makes change.
Select is responsible to move the coins to the proper buckets and put
the purchased item in the tray if purchase suceeds.

Arrays refactored to use named keys.

Type hinting used where applicable.

HOF array_reduce and array_filter used to avoid foreach loops.

Tests revised to use new Coin data structure.